+++ fe/src/Challenge/ChallengeView.test.js
@@ -1,9 +1,10 @@
+import { render, screen } from "@testing-library/react";
+import { act } from "react-dom/test-utils";
+
 import { Challenge } from "./Challenge";
 import { ChallengeView } from "./ChallengeView";
-
-import { render, screen } from "@testing-library/react";
-
 import { RootStoreProvider, initRootStoreAndSocket } from "../Store/RootStore";
+import { SoundBridge } from "../Sound/SoundBridge";
 
 import { NEW_AMAZONS_CHALLENGE_DATA } from "./Challenge.test";
 import { AMAZONS_CHALLENGE_DATA_AFTER_MOVE } from "./Challenge.test";
@@ -11,6 +12,28 @@ import { AMAZONS_CHALLENGE_DATA_AFTER_MOVE } from "./Challenge.test";
 const createFreshRootStore = () => {
   const RootStore = initRootStoreAndSocket();
   return RootStore;
+};
+
+const createFreshAmazonsChallenge = () => {
+  return new Challenge(NEW_AMAZONS_CHALLENGE_DATA);
+};
+
+const createAmazonsChallengeWithOneMove = () => {
+  return new Challenge(AMAZONS_CHALLENGE_DATA_AFTER_MOVE);
+};
+
+const renderChallenge = (challenge, RootStore = null) => {
+  if (RootStore === null) {
+    RootStore = createFreshRootStore();
+  }
+
+  const { asFragment } = render(
+    <RootStoreProvider store={RootStore}>
+      <ChallengeView challenge={challenge}></ChallengeView>
+    </RootStoreProvider>
+  );
+
+  return asFragment();
 };
 
 describe("ChallengeView", () => {
@@ -22,28 +45,61 @@ describe("ChallengeView", () => {
   });
 
   it("renders challenge with challenge data", () => {
-    const fresh_challenge = new Challenge(NEW_AMAZONS_CHALLENGE_DATA);
-    const RootStore = createFreshRootStore();
-
-    const { asFragment } = render(
-      <RootStoreProvider store={RootStore}>
-        <ChallengeView challenge={fresh_challenge}></ChallengeView>
-      </RootStoreProvider>
-    );
-    expect(asFragment()).toMatchSnapshot();
+    let chall_view = renderChallenge(createFreshAmazonsChallenge());
+    expect(chall_view).toMatchSnapshot();
   });
 
   it("renders challenge with at least one move and the last move indicator", () => {
-    const challenge_with_one_move = new Challenge(
-      AMAZONS_CHALLENGE_DATA_AFTER_MOVE
-    );
-    const RootStore = createFreshRootStore();
+    let chall_view = renderChallenge(createAmazonsChallengeWithOneMove());
+    expect(chall_view).toMatchSnapshot();
+  });
+});
 
-    const { asFragment } = render(
-      <RootStoreProvider store={RootStore}>
-        <ChallengeView challenge={challenge_with_one_move}></ChallengeView>
-      </RootStoreProvider>
-    );
-    expect(asFragment()).toMatchSnapshot();
+describe("new move sound effect", () => {
+  beforeEach(() => {
+    SoundBridge.mockClear();
+  });
+
+  it("plays a sound when challenge is updated with a new move", () => {
+    const challenge = createFreshAmazonsChallenge();
+    renderChallenge(challenge);
+
+    act(() => {
+      // makes sure all updates are flushed before moving to the next line
+      challenge.update_challenge_information(AMAZONS_CHALLENGE_DATA_AFTER_MOVE);
+    });
+
+    expect(
+      SoundBridge.mock.instances[0].playMoveSoundEffect.mock.calls.length
+    ).toBe(1);
+  });
+
+  describe("is not triggered when", () => {
+    afterEach(() => {
+      expect(
+        SoundBridge.mock.instances[0].playMoveSoundEffect
+      ).not.toHaveBeenCalled();
+    });
+
+    it("initial, fresh render", () => {
+      renderChallenge(createFreshAmazonsChallenge());
+    });
+
+    it("initial render of a in-progress game with moves", () => {
+      renderChallenge(createAmazonsChallengeWithOneMove());
+    });
+
+    it("challenge is updated without a move", () => {
+      const challenge = createFreshAmazonsChallenge();
+      renderChallenge(challenge);
+
+      act(() => {
+        // makes sure all updates are flushed before moving to the next line
+        challenge.update_challenge_information({
+          ...NEW_AMAZONS_CHALLENGE_DATA,
+          status: "OVER_DISCONNECT",
+        });
+      });
+    });
   });
 });